Added branches subcommand
Added the branches subcommand which will print out all projects that
have a topic branch in their repo. This makes it very easy to find
which repos have what topic branches and pairs with the checkout
subcommand which allows developers to easily switch between their
cross repo changes and check on the status of these changes.